# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a service to return questions that when answered will return elegible benefits.",
        "title": "Question Prioritization Service",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/healthcheck": {
            "get": {
                "description": "Returns Healthy",
                "summary": "Returns Healthy",
                "operationId": "healthcheck",
                "responses": {
                    "200": {
                        "description": "Healthy",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/next": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Request Prioritized Questions",
                "operationId": "next-question",
                "parameters": [
                    {
                        "description": "value",
                        "name": "Key",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bindings.NextQuestionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/renderings.NextQuestionResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "bindings.NextQuestionRequest": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                }
            }
        },
        "renderings.NextQuestionResponse": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        }
    }
}